--- a/backend/segmentation/api/main.py
+++ b/backend/segmentation/api/main.py
@@ -22,6 +22,7 @@ from contextlib import asynccontextmanager
 from api.routes import router
 from api.models import ErrorResponse, HealthResponse
 from api.metrics_endpoint import router as metrics_router
+from api.monitoring import router as monitoring_router
 from ml.model_loader import ModelLoader
 
 # Configure logging
@@ -95,6 +96,7 @@ app.add_middleware(
 # Include routes
 app.include_router(router, prefix="/api/v1")
 app.include_router(metrics_router)
+app.include_router(monitoring_router, prefix="/api/v1")
 
 # Global exception handler
 @app.exception_handler(Exception)