… .serena/memories/gpu_optimization_complete.md
… .serena/memories/green_deployment_fixes_2025-08-27.md
… .serena/memories/pr12_gpu_optimization_review.md
… .serena/memories/security_fixes_session_2025-08-27.md
… .serena/memories/security_patterns_and_solutions.md
… .serena/memories/session_email_config_2025-08-27.md
… .serena/memories/session_pr12_complete_2025-01-27.md
… .serena/memories/share_link_fix_2025-08-27.md
… .serena/memories/upload_issues_fix_2025-08-27.md
… NGINX_UNIFICATION_SUMMARY.md
… SENDGRID_DNS_SETUP.md
… backend/segmentation/config/batch_sizes.json
… backend/segmentation/monitoring/gpu_monitor.py
… backend/src/services/emailRetryService.ts
… backend/src/services/sharingService.ts
… backend/src/utils/envValidator.ts
… docker-compose.green.gpu.yml
… docker-compose.green.yml
… docker-compose.staging.yml
… docker-compose.yml
… docker/ml-cuda12.Dockerfile
… docker/nginx/nginx.prod.conf
… docs/AUTOMATIC_CLEANUP.md
… docs/GPU-CONFIGURATION.md
… docs/GPU-FALLBACK-BEHAVIOR.md
… docs/GPU-STATUS.md
… docs/batch-optimization-results.md
… scripts/apply-upload-fixes.sh
… scripts/auto-cleanup-disk.sh
… scripts/batch-optimization.py