## Summary
This PR brings GPU optimization for ML segmentation models and production environment configuration updates.

## Changes

### 🚀 GPU Optimization
- Enabled GPU support with CUDA 12.4 for ML segmentation
- Configured batch processing with dynamic memory management
- Achieved 17-34x performance improvements:
  - HRNet: 3.1s → 0.18s per image
  - ResUNet Small: 6.9s → 0.20s per image  
  - ResUNet Advanced: 18.1s → 1.2s per image

### 📧 Email Configuration
- Updated SMTP configuration for UTIA mail server
- Increased connection timeouts from 30s to 60s
- Added comprehensive debug logging for troubleshooting

### 🔧 Production Fixes
- Fixed nginx routing for green environment
- Updated translation files for all supported languages
- Added GPU documentation and benchmarking scripts

## Files Changed
- GPU Docker configurations and Dockerfiles
- Batch size optimization configuration
- Email service timeout adjustments
- Production environment routing fixes
- Comprehensive documentation for GPU setup

## Testing
- ✅ GPU performance benchmarked on RTX A5000
- ✅ Batch processing validated for all models
- ✅ Production site verified working
- ✅ All translations updated

## Documentation
- Added GPU configuration guide
- Created batch optimization results documentation
- Preserved session memories for configuration attempts

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- New Features
  - GPU-accelerated segmentation with batching delivering large speedups (reported up to ~78x).
  - API responses include timing, device and batch metadata for inference.
  - Share-link handling improved to reliably assign shared projects after sign-in/registration.

- Bug Fixes
  - Upload reliability improved (larger timeouts, directory/permission fixes).

- Documentation
  - New GPU setup/status, batch-optimization, and deployment/runbooks added.

- Tests
  - Multiple GPU/benchmarking and batch-verification scripts added.

- Chores
  - Email sending made configurable and more resilient; env-based deployment configs and example env added; routing switched to green.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->